Added checker for user already existance.

+++ utils/logger.js
@@ -7,11 +7,30 @@ const saveUserData = (userData) => {
     let data = [];
 
     if (fs.existsSync(usersFilePath)) {
-        const raw = fs.readFileSync(usersFilePath);
-        data = JSON.parse(raw);
+        const raw = fs.readFileSync(usersFilePath, "utf-8");
+        if (raw.trim()) {
+            data = JSON.parse(raw);
+        }
     }
+
+    // CHECK USER IS ALREADY EXISTS OR NOT.
+    const existingUser = data.find(
+        (user) => user.username.toLowerCase() === userData.username.toLowerCase()
+    );
+    if (existingUser) {
+        return {
+            success: false,
+            message: "User already exists."
+        }
+    }
+
+    // IF USER NOT EXISTS THEN SAVE THE USER DATA TO THE FILE.
     data.push(userData);
     fs.writeFileSync(usersFilePath, JSON.stringify(data, null, 2));
+    return {
+        success: true,
+        message: "User created successfully."
+    }
 };
